Add static method to allow for CefSharp.Wpf.HwndHost to share the same code.
